fastapi_backend/requirements.txt
fastapi
openai
requests
uvicorn
python-dotenv
SQLAlchemy
psycopg2-binary
alembic